Split components into containers and conmponents
