+++ src/containers/task-list/task-list.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Layout from '../main/layout.jsx';
+import Layout from '../../containers/layout/layout.jsx';
 
 import NavigationDrawer from '../navigation-drawer/navigation-drawer.jsx';
 
@@ -8,12 +8,6 @@ import { getProjects, getTasks, switchContext } from '../../actions';
 
 import { connect } from 'react-redux';
 
-import lightTheme from '../../themes/light.js';
-
-import themeManager from 'material-ui/lib/styles/theme-manager';
-import themeDecorator from 'material-ui/lib/styles/theme-decorator';
-
-@themeDecorator(themeManager.getMuiTheme(lightTheme))
 @connect(state => ({
   projects: state.projects.list,
   context: state.application.context,
